handson-app/requirements.txt
azure-cosmos
azure-functions
azure-search-documents==11.4.0
openai==1.3.*
